Add support for images in signature/image fields.
Whether a signature field expects a digital signature or a text signature, an image can now be added into that field from a file or a base64 encoded string. In fact, adding images works for any field, and the image will always overwrite whatever was in the field while automatically resizing to fill the bounds.