Change the path for temporary files to use the OS default.
Previously, all temporary files would be briefly created within the same directory as the original PDF file and destroyed shortly after. This did not work on read-only file systems, such as the ones provided by AWS lambda. Now all temporary files are stored in a directory that is defined by the operating system, which is writeable in AWS Lambda.